1) Smallest Missing Positive Integer -- algorithm Write a function:
class Solution { public int solution(int[] A); }

that, given an array A of N integers, returns the smallest positive integer (greater than 0) that does not occur in A.

For example, given A = [1, 3, 6, 4, 1, 2], the function should return 5.

Given A = [1, 2, 3], the function should return 4.

Given A = [−1, −3], the function should return 1.

Write an efficient algorithm for the following assumptions:

N is an integer within the range [1..100,000];
each element of array A is an integer within the range [−1,000,000..1,000,000].

2) PermMissingElem - algorithm

You are given an array A of N integers, where N is a positive integer. The array contains integers in the range from 1 to N + 1. However, one element is missing from the array, and your task is to find that missing element.

To clarify, the array A is guaranteed to contain N distinct integers. The missing element is an integer that should be in the range from 1 to N + 1, but it is not present in the array.

Your goal is to implement a function solution(int[] A) that takes the array A as input and returns the value of the missing element.

For example, given an array A with N = 4 such that:
A[0] = 2
A[1] = 3
A[2] = 1
A[3] = 5

The function should return 4, as it is the missing element in the range from 1 to N + 1.

The function should handle efficient algorithms for cases where N can vary and can be as large as 100,000. Additionally, the elements of the array A are all distinct and within the range from 1 to N + 1.